refactor templateProcessService
